merge in release v2.0.0

+++ src/TX/Config/ANT_ConfigureAdvancedBurst.cpp
@@ -89,8 +89,9 @@ uint8_t ConfigureAdvancedBurst::getDataLength() {
 
 #ifdef NATIVE_API_AVAILABLE
 
-uint8_t ConfigureAdvancedBurst::execute() {
-    // TODO
+uint32_t ConfigureAdvancedBurst::execute() {
+    // TODO arguments need to be packed into an array
+    // return sd_ant_adv_burst_config_set();
     return 0;
 }
